# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/plugins/install": {
      "post": {
        "operationId": "PluginsController_install",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Plugin installed"
          },
          "400": {
            "description": "Invalid package"
          },
          "409": {
            "description": "Plugin already installed"
          }
        },
        "summary": "Install a plugin from an uploaded .zip package",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/install-url": {
      "post": {
        "operationId": "PluginsController_installFromUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallFromUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plugin installed"
          },
          "400": {
            "description": "Invalid URL, download failed, or invalid package"
          },
          "409": {
            "description": "Plugin already installed"
          }
        },
        "summary": "Install a plugin by downloading its .zip from a URL (SSRF-guarded)",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/catalog": {
      "get": {
        "operationId": "PluginsController_catalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Catalog entries"
          },
          "400": {
            "description": "Catalog could not be fetched or parsed"
          }
        },
        "summary": "List the remote plugin catalog, annotated with install state",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}": {
      "get": {
        "operationId": "PluginsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin details"
          },
          "404": {
            "description": "Plugin not found"
          }
        },
        "summary": "Get plugin by ID",
        "tags": [
          "plugins"
        ]
      },
      "delete": {
        "operationId": "PluginsController_uninstall",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin uninstalled"
          },
          "400": {
            "description": "Cannot uninstall (e.g. built-in)"
          },
          "404": {
            "description": "Plugin not found"
          }
        },
        "summary": "Uninstall a plugin (removes its files; built-ins are protected)",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}/enable": {
      "post": {
        "operationId": "PluginsController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin enabled successfully"
          }
        },
        "summary": "Enable a plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}/disable": {
      "post": {
        "operationId": "PluginsController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin disabled successfully"
          }
        },
        "summary": "Disable a plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}/config": {
      "put": {
        "operationId": "PluginsController_updateConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plugin configuration updated"
          }
        },
        "summary": "Update plugin configuration",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}/config-ui": {
      "get": {
        "operationId": "PluginsController_getConfigUi",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Config UI HTML"
          },
          "404": {
            "description": "Plugin not found or has no config UI"
          }
        },
        "summary": "Serve a plugin's sandboxed config-UI entry HTML (for an iframe srcdoc)",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}/config/{sessionId}": {
      "put": {
        "operationId": "PluginsController_updateSessionConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-session plugin configuration updated"
          },
          "400": {
            "description": "Plugin is global (not session-scoped)"
          },
          "404": {
            "description": "Plugin not found"
          }
        },
        "summary": "Set a plugin config override for a specific session (empty = clear it)",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}/sessions": {
      "put": {
        "operationId": "PluginsController_updateSessions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginSessionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plugin session activation updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginDto"
                }
              }
            }
          },
          "400": {
            "description": "Plugin is global (not session-scoped)"
          },
          "403": {
            "description": "A session-restricted key may not replace the full active set — full activation replacement requires an unrestricted key"
          },
          "404": {
            "description": "Plugin not found"
          }
        },
        "summary": "Set which sessions a session-scoped plugin is activated for (['*'] = all)",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/plugins/{id}/update": {
      "post": {
        "operationId": "PluginsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallFromUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plugin updated"
          },
          "400": {
            "description": "Invalid URL/package, id mismatch, or built-in"
          },
          "404": {
            "description": "Plugin not found"
          }
        },
        "summary": "Update an installed plugin in place from a URL (preserves config + enabled state)",
        "tags": [
          "plugins"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.14.2",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "This instance (the origin serving these docs)"
    },
    {
      "url": "http://{host}:{port}",
      "description": "Another instance (set host and port)",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "InstallFromUrlDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "HTTPS URL of the plugin .zip to download and install. Plain http:// is rejected: the package is executable code, so it must be integrity-protected in transit (hosts on private networks remain subject to the SSRF guard). Optional content pinning: append `#sha256=<64 hex>` (fragment — never sent to the server; query params are ignored) to require the downloaded archive to match that digest; a mismatch fails the install."
          }
        },
        "required": [
          "url"
        ]
      },
      "PluginConfigDto": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "description": "Plugin configuration object"
          }
        },
        "required": [
          "config"
        ]
      },
      "PluginSessionsDto": {
        "type": "object",
        "properties": {
          "sessions": {
            "description": "Sessions to activate the plugin for; ['*'] = all numbers, [] = none",
            "example": [
              "*"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "sessions"
        ]
      },
      "PluginDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Plugin ID"
          },
          "name": {
            "type": "string",
            "description": "Plugin name"
          },
          "version": {
            "type": "string",
            "description": "Plugin version"
          },
          "type": {
            "type": "string",
            "enum": [
              "engine",
              "storage",
              "queue",
              "auth",
              "extension"
            ],
            "description": "Plugin type"
          },
          "description": {
            "type": "string",
            "description": "Plugin description"
          },
          "author": {
            "type": "string",
            "description": "Plugin author"
          },
          "status": {
            "type": "string",
            "enum": [
              "installed",
              "enabled",
              "disabled",
              "error"
            ],
            "description": "Plugin status"
          },
          "config": {
            "type": "object",
            "description": "Plugin configuration"
          },
          "builtIn": {
            "type": "boolean",
            "description": "Whether this is a built-in plugin"
          },
          "provides": {
            "description": "Features provided by this plugin",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ingressCapable": {
            "type": "boolean",
            "description": "Whether this plugin can host provisioned ingress instances"
          },
          "sessionScoped": {
            "type": "boolean",
            "description": "Whether the plugin is scoped to specific sessions (false = global)"
          },
          "activeSessions": {
            "description": "Sessions this plugin is activated for; ['*'] = all numbers",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "configSchema": {
            "type": "object",
            "description": "Configuration schema"
          },
          "configUi": {
            "type": "object",
            "description": "Sandboxed-iframe config editor (entry HTML + optional height)"
          },
          "i18n": {
            "type": "object",
            "description": "Localized dashboard text (name/description/config titles) per locale code"
          },
          "sessionConfig": {
            "type": "object",
            "description": "Per-session config overrides, keyed by sessionId (secrets redacted)"
          },
          "loadedAt": {
            "type": "string",
            "description": "When the plugin was loaded"
          },
          "enabledAt": {
            "type": "string",
            "description": "When the plugin was enabled"
          },
          "error": {
            "type": "string",
            "description": "Error message if plugin is in error state"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "type",
          "status",
          "config",
          "builtIn",
          "provides",
          "ingressCapable",
          "sessionScoped",
          "activeSessions"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}